rename(test) : 패키지 이동 및 이름 변경
