+++ src/test/java/kitchenpos/table/application/TableGroupServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.table.application;
 
 import static common.MenuGroupFixture.메뉴그룹_한마리;
 import static common.OrderTableFixture.from;
